relocate KEY_RUN_SIZE; auto-update run metadata on status change events
